separated out custom middleware from the express app - SOHO-7803 #4 #6
